+++ backend/main/models/__init__.py
@@ -1,0 +1,5 @@
+from .users import Users as UsersModel
+from .authors import Authors as AuthorsModel
+from .books import Books as BooksModel
+from .loans import Loans as LoansModel
+from .ratings import Ratings as RatingsModel